# api-nestjs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/update": {
      "patch": {
        "operationId": "updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário atualizado."
                    },
                    "updatedFields": {
                      "type": "object",
                      "example": {
                        "fisrtName": "Novo",
                        "lastName": "Nome",
                        "whatsappNumber": "11999999999"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar usuário atual",
        "tags": [
          "Usuários"
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Link de acesso enviado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Link de acesso enviado."
                    },
                    "verifyLoginUrl": {
                      "type": "string",
                      "example": "http://seusite.com/verify-login?token=xyz"
                    },
                    "loginToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          },
          "429": {
            "description": "Muitas tentativas, aguarde um momento."
          }
        },
        "summary": "Enviar magic link para login",
        "tags": [
          "Autenticação"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "registerUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Link de verificação enviado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Link de verificação enviado."
                    },
                    "registerToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "verifyRegisterUrl": {
                      "type": "string",
                      "example": "http://seusite.com/verify-register?token=xyz"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          },
          "429": {
            "description": "Muitas tentativas, aguarde um momento."
          }
        },
        "summary": "Criar novo usuário",
        "tags": [
          "Autenticação"
        ]
      }
    },
    "/auth/verify-login": {
      "get": {
        "operationId": "verifyLogin",
        "parameters": [
          {
            "name": "loginToken",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login realizado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Login realizado."
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado."
          },
          "429": {
            "description": "Muitas tentativas, aguarde um momento."
          }
        },
        "summary": "Verificar magic link",
        "tags": [
          "Autenticação"
        ]
      }
    },
    "/auth/verify-register": {
      "get": {
        "operationId": "verifyRegister",
        "parameters": [
          {
            "name": "registerToken",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário verificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário verificado."
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado."
          },
          "429": {
            "description": "Muitas tentativas, aguarde um momento."
          }
        },
        "summary": "Verificar registro de usuário",
        "tags": [
          "Autenticação"
        ]
      }
    },
...
    "/auth/refresh-token": {
      "get": {
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "refreshToken",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token atualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token atualizado."
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado."
          },
          "429": {
            "description": "Muitas tentativas, aguarde um momento."
          }
        },
        "summary": "Renovar token de acesso",
        "tags": [
          "Autenticação"
        ]
      }
    }
  },
  "info": {
    "title": "API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "usuario@email.com"
          },
          "firstName": {
            "type": "string",
            "description": "Primeiro nome do usuário",
            "example": "Nome"
          },
          "lastName": {
            "type": "string",
            "description": "Sobrenome do usuário",
            "example": "Sobrenome"
          },
          "whatsappNumber": {
            "type": "string",
            "description": "Número do WhatsApp",
            "example": "11999999999"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "usuario@email.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "usuario@email.com"
          },
          "firstName": {
            "type": "string",
            "description": "Primeiro nome do usuário",
            "example": "Nome"
          },
          "lastName": {
            "type": "string",
            "description": "Sobrenome do usuário",
            "example": "Sobrenome"
          },
          "whatsappNumber": {
            "type": "string",
            "description": "Número do WhatsApp",
            "example": "11999999999"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName"
        ]
      }
    }
  }
}